+++ Changelog.md
@@ -1,5 +1,9 @@
 # Change Log
 
+#### 0003
+
+- Edge case bug causing update to fail resolves - requires Simple Calendar v1.3.66 or greater
+
 #### 0002
 
 - 0.8.6 compatibility confirmed